backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
django-cors-headers==4.3.1
# Add more later for tools
rdkit==2023.9.6 #for molecule parsing and reactions 
pubchempy==1.0.4 #for compound metadata 
transformers==4.41.2 #for loading ChemBerta
torch==2.3.0 #Required for transformers cpu version
pytest==8.3.2
pytest-django==4.9.0
llama-index==0.10.0
faiss-cpu==1.8.0
pypdf==4.3.0
llama-index-vector-stores-faiss==0.1.0
llama-index-embeddings-huggingface==0.1.0
langchain==0.2.0

